organized functions for category filter
