+++ js/handler/handleCategoryButtons.mjs
@@ -1,0 +1,52 @@
+import { filterPostsByCategory, loadMoreWrapper } from "../handler/handlePosts.mjs";
+
+// highlight the category buttons and update headline to category name
+export const handleCategoryButtons = () => {
+  const buttons = document.querySelectorAll("#category-button");
+  const headline = document.querySelector(".main-heading");
+  const allPostsButton = document.querySelector(".all-posts-btn.all-posts-btn");
+  // call the function with categoryId for all posts initially to display all posts and highlight the "All" category button.
+  filterPostsByCategory("31");
+  // highlight the "All" button initially
+  allPostsButton.classList.add("active");
+
+  buttons.forEach((button) => {
+    button.addEventListener("click", (event) => {
+      // Get the category ID from the data attribute
+      const categoryId = button.getAttribute("category-id");
+      // change the h1 to display the category name
+      headline.textContent = event.target.textContent;
+
+      // Check if the category buttons already has the class "active"
+      if (button.classList.contains("active")) {
+        // If it has, remove the class "active" (styling)
+        button.classList.remove("active");
+        // and add class to "All" category button.
+        allPostsButton.classList.add("active");
+        // change the h1 to display the category name
+        headline.textContent = "All articles";
+        // call the function to reset to show all posts if no category is active
+        filterPostsByCategory("31");
+        // display the "load more" button when category button is clicked twice(deactivated)
+        loadMoreWrapper.classList.remove("is-hidden");
+      } else {
+        // Remove active class from all buttons
+        buttons.forEach((btn) => btn.classList.remove("active"));
+        // Add class "active" to clicked button
+        button.classList.add("active");
+        // and remove from "All" category button.
+        allPostsButton.classList.remove("active");
+        // to get the "All" category button display the "load more" button on the bottom
+        if (categoryId === "31") {
+          loadMoreWrapper.classList.remove("is-hidden");
+          // highlight the "All" button when clicked
+          allPostsButton.classList.add("active");
+          // change the h1 to display the category name
+          headline.textContent = "All articles";
+        }
+        // Call the function to filter posts by the selected category
+        filterPostsByCategory(categoryId);
+      }
+    });
+  });
+};
